Update layout of splatnet assets in battle template
The positional attributes for the splatnet assets in the battle_opengraph.html were adjusted to improve layout. The 'squid_tape.html' component was re-positioned (moved slightly to the left and upward) and its scale was increased. This change ensures that the component is properly displayed within the 'battle' page under 'FEST' mode.