Add hover effect and verification icon to profile URLs
In the profile_sidebar.html, added a hover underline effect to the URLs. This will give clearer indication that these are clickable links. Also, an SVG verification icon is added next to verified URLs. This provides visual confirmation of the verification status of a URL, enhancing the user experience. No changes were made to the poetry.lock file in this commit.